chore(ui): Rename `get-run-duration` to `calculate-duration`
There is no "getter" function in there, and durations are not restricted
to runs.
