+++ ui/src/helpers/calculate-duration.ts
